+++ Source/EventFlow.ReadStores.Elasticsearch/IElasticsearchReadModelStore.cs
@@ -20,20 +20,10 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System.Collections.Generic;
-using EventFlow.Aggregates;
-using EventFlow.Specifications;
-
-namespace EventFlow.Provided.Specifications
+namespace EventFlow.ReadStores.Elasticsearch
 {
-    public class AggregateIsNewSpecification : Specification<IAggregateRoot>
+    public interface IElasticsearchReadModelStore<TReadModel> : IReadModelStore<TReadModel>
+        where TReadModel : class, IReadModel, new()
     {
-        protected override IEnumerable<string> IsNotSatisfiedBecause(IAggregateRoot obj)
-        {
-            if (!obj.IsNew)
-            {
-                yield return $"'{obj.Name}' with ID '{obj.GetIdentity()}' is not new";
-            }
-        }
     }
 }