+++ Source/EventFlow.ReadStores.Elasticsearch/ReadModelDescription.cs
@@ -20,20 +20,20 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System.Collections.Generic;
-using EventFlow.Aggregates;
-using EventFlow.Specifications;
+using System;
 
-namespace EventFlow.Provided.Specifications
+namespace EventFlow.ReadStores.Elasticsearch
 {
-    public class AggregateIsNewSpecification : Specification<IAggregateRoot>
+    public class ReadModelDescription
     {
-        protected override IEnumerable<string> IsNotSatisfiedBecause(IAggregateRoot obj)
+        public ReadModelDescription(
+            IndexName indexName)
         {
-            if (!obj.IsNew)
-            {
-                yield return $"'{obj.Name}' with ID '{obj.GetIdentity()}' is not new";
-            }
+            if (indexName == null) throw new ArgumentNullException(nameof(indexName));
+
+            IndexName = indexName;
         }
+
+        public IndexName IndexName { get; }
     }
 }